# Cold start repro and optimization