--- a/lib/api/src/sys/module.rs
+++ b/lib/api/src/sys/module.rs
@@ -268,20 +268,27 @@ impl Module {
         resolver: &dyn Resolver,
     ) -> Result<InstanceHandle, InstantiationError> {
         unsafe {
-            let instance_handle = self.artifact.instantiate(
-                self.store.tunables(),
-                resolver,
-                Box::new((self.store.clone(), self.artifact.clone())),
-            )?;
+            let instance_handle = {
+                let _span = tracing::debug_span!(target: "vm", "artifact.instantiate").entered();
+                self.artifact.instantiate(
+                    self.store.tunables(),
+                    resolver,
+                    Box::new((self.store.clone(), self.artifact.clone())),
+                )?
+            };
 
             // After the instance handle is created, we need to initialize
             // the data, call the start function and so. However, if any
             // of this steps traps, we still need to keep the instance alive
             // as some of the Instance elements may have placed in other
             // instance tables.
-            self.artifact
-                .finish_instantiation(&self.store, &instance_handle)?;
 
+            {
+                let _span =
+                    tracing::debug_span!(target: "vm", "artifact.finish_instantiation").entered();
+                self.artifact
+                    .finish_instantiation(&self.store, &instance_handle)?;
+            }
             Ok(instance_handle)
         }
     }