@Longarithm found a [many function contracts](https://github.com/near/nearcore/compare/90e182b6..8814b7c0f) which compile and execute pretty slow. This PR attempt to address the **execute** part only.

@Longarithm's result:
```
contract size = 1456KiB
vm_kind = Wasmer0
time to run no-op fn = 2.970571573s
time to run no-op fn = 640.596788ms
time to run no-op fn = 636.828433ms
time to run no-op fn = 634.383194ms
time to run no-op fn = 634.292482ms
vm_kind = Wasmer2
time to run no-op fn = 6.357417752s
time to run no-op fn = 186.123842ms
time to run no-op fn = 185.622875ms
time to run no-op fn = 190.486191ms
time to run no-op fn = 187.222724ms
```

Note the first time of wasmer0/wasmer2 including compilation + execution, and subsequent calls including execution only.

My local is about 4x faster, which I'm using as a baseline (46ms for wasmer 2 init, correspond to above 187ms). And in this PR, I move the bench in wasmer repo and it can be run to check bottlenecks and the degree of optimization. It can be run by:
```
cargo run --release --package wasmer-workspace --example compiler-singlepass --features singlepass
```

After current version of this PR, it reduce initialization time from 46ms to 14ms. Further optimization maybe possible, I'll update the number as i progress. The PR is in draft because of tracing debug prints and one error checking TODO